+++ rust/lance/src/dataset/write/merge_insert/exec/indexed_lookup.rs
